vhost: extract socket I/O into transport
The core vhost-user protocol code should not do socket I/O, because the
details are transport-specific.  Move code to send and receive
vhost-user messages into trans_af_unix.c.

The connection fd is a transport-specific feature. Therefore, it should
and eventually will be removed from the core vhost-user code. That is,
it will be removed from the vhost_user_msg_handler() and the message
handlers. We keep it for now, because vhost_user_set_mem_table() needs
it. In a later commit, we will refactor the map/unmap functionality and
after that we will be able to remove the connection fds from the core
vhost-user code.

Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>